+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Load the icon fonts as soon as the editor starts
 * Make the `rerender` method async
 * Use `child_class` property in the populate method
+* Add filter logic to widgets for getting specific values
 
 ## 0.1.5 (2022-07-14)